Add User model to Prisma schema
This commit adds a new User model to the Prisma schema. The User model includes fields for id, name, email, password, and created_at. The id field is set to autoincrement, the email field is marked as unique, and the created_at field is set to the current date and time by default.

The purpose of this change is to create a User model that can be used to store user information in the database.