Add migration_lock.toml file for version control
This commit adds the migration_lock.toml file, which should not be manually edited and should be added to the version-control system (i.e. Git). The file specifies the provider as "postgresql".

The purpose of this change is to include the migration_lock.toml file in the repository to ensure proper version control of the database migrations.