Add createUser endpoint to UserController
This commit adds the `createUser` endpoint to the `UserController`. The endpoint handles the creation of a new user by checking if the email is already taken and creating a new user if it is not. It returns a success message and the newly created user data.

The purpose of this change is to provide functionality for creating new users in the system.