Increment comment count when creating a comment and decrement when deleting
This commit adds code to the `CommentController.js` file to increment the comment count of a post when a comment is created and decrement the comment count when a comment is deleted. This enhances the functionality of the application by keeping track of the number of comments associated with each post.