Add searchPost endpoint for searching posts by description
This commit adds a new endpoint `searchPost` to the `PostController.js` file. The endpoint allows users to search for posts based on the description using a query parameter. The search is performed using the `prisma.post.findMany` method with a `search` condition on the `description` field. This change enhances the functionality of the application by providing a way to search for specific posts based on their description.