Add cache strategy to fetch all posts
This commit adds a cache strategy to the `findMany` method in the `onlyShowPosts` function of the `PostController.js` file. The cache strategy includes a `swr` value of 60 seconds and a `ttl` value of 60 seconds. This change improves the performance of the application by caching the posts data and reducing unnecessary database queries.